圧力センサを押したときにsummarize.pyを起動させることに成功
